Put staking slice in its own directory and extract a helper
